refactor(Test): 파일 이동으로인한 오류 수정
