+++ common/src/test/java/kitchenpos/common/testassistance/config/DBCleaner.java
@@ -56,6 +56,7 @@ public class DBCleaner implements InitializingBean {
         jdbcTemplate.execute("INSERT INTO order_table (id, number_of_guests, empty) VALUES (6, 0, true)");
         jdbcTemplate.execute("INSERT INTO order_table (id, number_of_guests, empty) VALUES (7, 0, true)");
         jdbcTemplate.execute("INSERT INTO order_table (id, number_of_guests, empty) VALUES (8, 0, true)");
+        jdbcTemplate.execute("INSERT INTO order_table (id, number_of_guests, empty) VALUES (9, 10, false)");
     }
 
     @Override
