+++ table/src/main/java/kitchenpos/table/application/ValidateEmptyTableHandler.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Component;
 
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.orders.event.ValidateEmptyTableEvent;
-import kitchenpos.order.exception.NotFoundOrderException;
+import kitchenpos.common.event.ValidateEmptyTableEvent;
 import kitchenpos.table.exception.EmptyOrderTableException;
+import kitchenpos.table.exception.NotFoundOrderTableException;
 
 @Component
 public class ValidateEmptyTableHandler {
@@ -21,7 +21,7 @@ public class ValidateEmptyTableHandler {
 
     @EventListener
     public void handle(ValidateEmptyTableEvent event) {
-        OrderTable orderTable = orderTableRepository.findById(event.getOrderTableId()).orElseThrow(NotFoundOrderException::new);
+        OrderTable orderTable = orderTableRepository.findById(event.getOrderTableId()).orElseThrow(NotFoundOrderTableException::new);
 
         if (orderTable.isEmpty()) {
             throw new EmptyOrderTableException("주문테이블이 빈테이블입니다.");
